% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tree.R
\docType{methods}
\name{as_tree}
\alias{as_tree}
\alias{as_tree,phylo-method}
\alias{as_tree,}
\alias{phylo,}
\alias{phylo-method}
\alias{as_tree,phylo4-method}
\alias{as_tree,}
\alias{phylo4,}
\alias{phylo4-method}
\title{as_tree}
\usage{
as_tree(phy)

\S4method{as_tree}{phylo}(phy)

\S4method{as_tree}{phylo4}(phy)
}
\arguments{
\item{phy}{Required. phy is either a \code{\link[ape]{phylo}} class object,
\code{\link[phylobase]{phylo4}} class object, or a character that can be a newick
 literal or apath to newick-containing file.}
}
\description{
Convert objects to newick strings
}
