% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.branches.R
\name{collapse_branch}
\alias{collapse_branch}
\title{collapse branch}
\usage{
collapse_branch(phycnv, nodeid, collapse = TRUE)
}
\arguments{
\item{phycnv}{Required.}

\item{nodeid}{Required.}

\item{collapse}{Optional. Default \code{TRUE}.  Whether to collapse the node.}
}
\description{
given a \code{\link{phylocanvas}} object and an id node,
collapse the node.
}
\seealso{
\url{http://phylocanvas.org/docs/features/#collapse-branches}
}
