% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.branches.R
\name{highlight_node}
\alias{highlight_node}
\title{highlight node}
\usage{
highlight_node(phycnv, nodeid)
}
\arguments{
\item{phycnv}{Required.}

\item{nodeid}{Required.}
}
\description{
given a \code{\link{phylocanvas}} object and an nodeid,
highlight the node.
}
\seealso{
\url{http://phylocanvas.org/docs/features/#highlight-leaves}
}

