% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.branches.R
\name{prune_branch}
\alias{prune_branch}
\title{prune branch}
\usage{
prune_branch(phycnv, nodeid)
}
\arguments{
\item{phycnv}{Required.}

\item{nodeid}{Required.}
}
\description{
given a \code{\link{phylocanvas}} object and an nodeid for
an internal node, prune the branch.
}
\seealso{
\url{http://phylocanvas.org/docs/features/#prune-branches}
}

