% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.branches.R
\name{rotate_branch}
\alias{rotate_branch}
\title{rotate a branch}
\usage{
rotate_branch(phycnv, nodeid)
}
\arguments{
\item{phycnv}{Required.}

\item{nodeid}{Required.}
}
\description{
given a \code{\link{phylocanvas}} object and an nodeid for
an internal node, rotate the branch defined by the node
}
\seealso{
\url{http://phylocanvas.org/docs/features/#rotate-branches}
}
