% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.branches.R
\name{select_branch}
\alias{select_branch}
\title{select  branch}
\usage{
select_branch(phycnv, nodeid, cascade = FALSE)
}
\arguments{
\item{phycnv}{Required.}

\item{nodeid}{Required.}

\item{cascade}{Optional. Default \code{FALSE}}
}
\description{
given a \code{\link{phylocanvas}} object and an nodeid for
an internal node, select the branch defined by the node
}
\seealso{
\url{http://phylocanvas.org/docs/features/#select-branches}
}
