% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.branches.R
\name{style_node}
\alias{style_node}
\title{style node}
\usage{
style_node(phycnv, nodeid, highlighted = FALSE, color = "black",
  shape = "circle", nodesizeratio = 1, strokecolor = "black",
  fillcolor = "black", linewidth = 1, labelcolor = "black",
  labeltextsize = 20, labelfont = "Arial", labelformat = "bold")
}
\arguments{
\item{phycnv}{Required.}

\item{nodeid}{Required.}

\item{highlighted}{Optional. Default \code{FALSE}.  Whether to highlight the node.}

\item{color}{Optional. Color of the node/branch}

\item{shape}{Optional. Default \code{'circle'}. Can be either 'circle', 'square' or 'traingle'.}

\item{nodesizeratio}{Optional. Default \code{1}. Scale ratio of the nodesize to the base node size.}

\item{strokecolor}{Optional. Default \code{"black"}. Colors are strings in the format of names, rgb strings or hex values.}

\item{fillcolor}{Optional. Default \code{"black"}. Scale ratio of the node.}

\item{linewidth}{Optional. Default \code{1}.}

\item{labelcolor}{Optional. Default \code{"black"}. Label color. Colors are strings in the format of names, rgb strings or hex values.}

\item{labeltextsize}{Optional.Default \code{20}. Label size.}

\item{labelfont}{Optional. Default \code{"Arial"}. Label font.}

\item{labelformat}{Optional. Default \code{"bold"}. Label format.}
}
\description{
add styles to a node
}
\details{
given a \code{\link{phylocanvas}} object and nodeid, apply
the given styles to that node.
}
\seealso{
\url{http://phylocanvas.org/docs/features/#style-branches-individually}
}
